Separate Race Day Summary view on the Home View

+++ src/RaceDay.Core/Entities/RaceLapEntity.cs
@@ -9,6 +9,6 @@ public class RaceLapEntity : IEntity
     public Guid LapGuid { get; init; }
     public Guid RacerGuid { get; init; }
     public int LapNumber { get; set; }
-    public float LapTimeSeconds { get; set; }
+    public double LapTimeSeconds { get; set; }
     public float LapSpeed { get; set; }
 }
